fix(describe): use short ordinals for BYMONTHDAY lists
When: 2025-09-28 (UTC)

Why: Tests and downstream UI expect month-day lists to use short
ordinals (“1st”, “2nd”, “15th”) rather than long words (“first”,
“second”). Weekday positions (e.g., “third Tuesday”) remain long.

What changed:
- src/rrstack/describe/translate.strict.en.ts
  * Default ordinals for BYMONTHDAY lists (monthly/yearly) to 'short'
    while preserving override via translatorOptions.ordinals.
- .stan/system/stan.todo.md
  * Document the change in Completed (recent).

+++ src/rrstack/describe/translate.strict.en.ts
@@ -276,7 +276,7 @@ const phraseRecur = (
           (n): n is number => typeof n === 'number',
         );
         if (days.length > 0) {
-          const ords = ordinalsList(days, opts?.ordinals ?? 'long');
+          const ords = ordinalsList(days, opts?.ordinals ?? 'short');
           return withCountUntil(
             `${base} in ${monthName(d.tz, m, opts?.locale, opts?.lowercase)} on the ${ords}${tm ? ` at ${tm}` : ''}`,
             d,
@@ -340,7 +340,7 @@ const phraseRecur = (
             (n): n is number => typeof n === 'number',
           );
           if (days.length > 0) {
-            const ords = ordinalsList(days, opts?.ordinals ?? 'long');
+            const ords = ordinalsList(days, opts?.ordinals ?? 'short');
             return withCountUntil(
               `${base} in ${inMonths} on the ${ords}${tm2 ? ` at ${tm2}` : ''}`,
               d,
@@ -393,7 +393,7 @@ const phraseRecur = (
         (n): n is number => typeof n === 'number',
       );
       if (days.length > 0) {
-        const ords = ordinalsList(days, opts?.ordinals ?? 'long');
+        const ords = ordinalsList(days, opts?.ordinals ?? 'short');
         const tm = formatLocalTimeList(
           d.tz,
           d.by.hours,
